requirements.txt
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
streamlit
pypdf